fix: update file name as remote
